Adding mercurial and eclipse project configuration
This has historical reasons. Mercurial was the previous used code
repository. However, a switch was needed for a better branch support.

Eclipse was previously used, however, intellij idea has its pros and
cons. A big one is the fast and reliable JUnit and coverage support.
The maven integration and its dependency management helps in developing
fast and reliable software.

This commit has to be tagged if some information is missing. Maybe it
can be retrieved using the inner mercurial code repository.