Adding current project environment: intellij / git
intellij idea will be the supported integrated development environment.
A pre-defined module was added to. This module will address all core
classes. IDEA allows a flexibel handling of dependencies between modul-
es. Therefore classes will move from one module to another or back.

Git will be the code repository for the future. Branch support is
really needed and required for the next steps: this legacy system needs
a clean-up. Knowledge built up and for now only the SOAP interface has
to work properly again. After new interfaces can be supported. But for
now the actual behaviour has to be brought back to live.